pyproject.toml
[project]
    authors = [ { name = "Snaptraks" } ]
    dependencies = [
        "discord-py>=2.5.0",
        "pillow>=11.3.0",
        "snapcogs[all]",
        "sqlalchemy[asyncio]>=2.0.38",
    ]
    description = "Discord bot for the Hatventures Community server"
    name = "hatbot"
    readme = "README.md"
    requires-python = ">=3.12"
    version = "2025.02.1"

[tool.uv.sources]
    snapcogs = { git = "https://github.com/Snaptraks/SnapCogs.git", rev = "master" }

[tool.ruff.lint]
    ignore = [
        "D",       # docstrings
        "G004",    # logging-f-strings
        "PLR2004", # magic-value-comparison
        "S311",    # suspicious-non-cryptographic-random-usage
    ]
